+++ scripts/changelog.ts
@@ -63,6 +63,11 @@ async function exec(): Promise<void> {
         'Removed `state` prop from `Document`. You should pass only the id as `id` prop instead.',
         'Removed `value` property from `StateType.object`. The values are now exposed directly on `[key]`.'
       ]
+    },
+    {
+      added: ['GeoGebra plugin'],
+      changed: ['**plugin-spoiler**. Remove input element in render mode'],
+      fixed: ["**plugin-text**. Don't dispatch selection changes anymore"]
     }
   ])